refactor: Move ledger entry helper functions from View.h/View.cpp to dedicated helper files (#6453)
