+++ src/libxrpl/tx/transactors/dex/PermissionedDEXHelpers.cpp
@@ -1,4 +1,4 @@
-#include <xrpl/ledger/CredentialHelpers.h>
+#include <xrpl/ledger/helpers/CredentialHelpers.h>
 #include <xrpl/tx/transactors/dex/PermissionedDEXHelpers.h>
 
 namespace xrpl {
