+++ src/libxrpl/tx/transactors/lending/LoanBrokerCoverWithdraw.cpp
@@ -1,6 +1,6 @@
 #include <xrpl/tx/transactors/lending/LoanBrokerCoverWithdraw.h>
 //
-#include <xrpl/ledger/CredentialHelpers.h>
+#include <xrpl/ledger/helpers/CredentialHelpers.h>
 #include <xrpl/protocol/STTakesAsset.h>
 #include <xrpl/tx/transactors/lending/LendingHelpers.h>
 #include <xrpl/tx/transactors/payment/Payment.h>
