+++ src/test/app/Credentials_test.cpp
@@ -2,7 +2,7 @@
 
 #include <xrpl/basics/strHex.h>
 #include <xrpl/ledger/ApplyViewImpl.h>
-#include <xrpl/ledger/CredentialHelpers.h>
+#include <xrpl/ledger/helpers/CredentialHelpers.h>
 #include <xrpl/protocol/Feature.h>
 #include <xrpl/protocol/Indexes.h>
 #include <xrpl/protocol/Protocol.h>
