+++ src/xrpld/app/paths/detail/XRPEndpointStep.cpp
@@ -1,8 +1,8 @@
 #include <xrpld/app/paths/detail/StepChecks.h>
 
 #include <xrpl/basics/Log.h>
-#include <xrpl/ledger/Credit.h>
 #include <xrpl/ledger/PaymentSandbox.h>
+#include <xrpl/ledger/helpers/RippleStateHelpers.h>
 #include <xrpl/protocol/Feature.h>
 #include <xrpl/protocol/IOUAmount.h>
 #include <xrpl/protocol/Quality.h>
